+++ src/repositories/selects/shopSelect.js
@@ -1,5 +1,8 @@
 import { cardDetailSelect, cardSelect } from "./card-select.js";
-import { exchangeCardInfo, exchangeCardShopAndUserSelect } from "./exchange-select.js";
+import {
+  exchangeCardInfo,
+  exchangeCardShopAndUserSelect,
+} from "./exchange-select.js";
 
 export const shopSelect = {
   id: true,
