+++ src/services/exchange-service.js
@@ -7,6 +7,7 @@ import { exchangeMapper } from "../controllers/mappers/exchange-mapper.js";
 
 import { EXCHANGE_VOLUME } from "../constants/exchange.js";
 import shopRepository from "../repositories/shopRepository.js";
+import { exchangeDelete } from "../utils/sellout-util.js";
 
 async function checkExchangeByUser(userId, shopId) {
   const filter = {
@@ -56,6 +57,7 @@ async function acceptByExchange(userId, exchangeId, reqBody) {
     shopId,
     exchangeCardId,
     buyerId,
+    shopDetailData,
     shopCardId,
     hasSellerExchangeCard,
     hasBuyershopCard,
@@ -120,6 +122,11 @@ async function acceptByExchange(userId, exchangeId, reqBody) {
       const delteeExchange = await exchangeRepository.deleteData({
         id: exchangeId,
       });
+
+      // 매진 시 관련 exchange 삭제
+      if (shopDetailData.remainingQuantity === 1) {
+        await exchangeDelete(shopDetailData);
+      }
 
       const responseMappeing = exchangeMapper(exchangeData);
